display: Define soong namespace
HALs are starting to move to blueprint files, which means that pathmap
ifeqs no longer work. This results in modules with the same name in
different platform HAL directories. Since these modules aren't used or
linked anywhere outside the HAL directories and a device's device.mk,
we can safely use a soong namespace to isolate the different HAL
directories' blueprints.

Change-Id: I6f0b228ed36a6feb48b1fecc5c6793ae94e70dcc
Signed-off-by: Yash-Garg <ben10.yashgarg@gmail.com>